create abstract lunar phase class
