+++ lib/zakuro/version/senmyou/monthly/lunar_phase.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../../output/logger'
+require_relative '../../../calculation/monthly/abstract_lunar_phase'
 
 require_relative '../const/remainder'
 
@@ -21,22 +21,9 @@ module Zakuro
       #
       # LunarPhase 月の位相
       #
-      class LunarPhase
+      class LunarPhase < Calculation::Monthly::AbstractLunarPhase
         # @return [Cycle::Remainder] 弦
         QUARTER = Const::Remainder::Solar::QUARTER
-
-        # @return [Output::Logger] ロガー
-        LOGGER = Output::Logger.new(location: 'lunar_phase')
-
-        # @return [Array<String>] 月内の弦
-        PHASE_INDEXES = %w[朔日 上弦 望月 下弦].freeze
-
-        # @return [Remainder]] 経
-        attr_reader :average_remainder
-        # @return [SolarTerm] 二十四節気（入定気）
-        attr_reader :solar_term
-        # @return [Integer] 弦
-        attr_reader :phase_index
 
         #
         # 初期化
@@ -44,84 +31,21 @@ module Zakuro
         # @param [Integer] western_year 西暦年
         #
         def initialize(western_year:)
-          # 経
-          @average_remainder = Origin::AverageNovember.get(western_year: western_year)
           # 天正閏余
           lunar_age = Origin::LunarAge.get(western_year: western_year)
-          # 入定気
-          @solar_location = Solar::Location.new(lunar_age: lunar_age)
-          # 入暦
-          @lunar_location = Lunar::Location.new(
-            western_year: western_year,
-            lunar_age: Cycle::LunarRemainder.new(total: 0).add!(lunar_age)
+
+          super(
+            quater: QUARTER,
+            average_remainder: Origin::AverageNovember.get(western_year: western_year),
+            solar_location: Solar::Location.new(lunar_age: lunar_age),
+            lunar_location: Lunar::Location.new(
+              western_year: western_year,
+              lunar_age: Cycle::LunarRemainder.new(total: 0).add!(lunar_age)
+            )
           )
-
-          # 弦
-          @phase_index = 0
-        end
-
-        #
-        # 次の弦に進める
-        #
-        # @return [Remainder] 定朔
-        #
-        def next_phase
-          adjusted = current_remainder
-
-          add_quarter_moon_size
-
-          adjusted
-        end
-
-        #
-        # 次の月に進める
-        # @note 進めた後の月の定朔ではなく、当月のものを返却する
-        #
-        # @return [Remainder] 当月初の定朔
-        #
-        def next_month
-          result = nil
-          PHASE_INDEXES.each_with_index do |_phase, index|
-            adjust = next_phase
-            result = adjust if index.zero?
-          end
-
-          result
         end
 
         private
-
-        #
-        # 次の弦に進める
-        #
-        # @return [Integer] 弦
-        #
-        def next_phase_index
-          @phase_index += 1
-          @phase_index = 0 if @phase_index >= PHASE_INDEXES.size
-          @phase_index
-        end
-
-        #
-        # 朔月（月初）であるか
-        #
-        # @return [True] 朔月である
-        # @return [False] 朔月ではない
-        #
-        def first_phase?
-          @phase_index.zero?
-        end
-
-        #
-        # 朔月のみログ出力する
-        #
-        # @param [String] messages メッセージ（可変長）
-        #
-        def debug(*messages)
-          return unless first_phase?
-
-          LOGGER.debug(*messages)
-        end
 
         # :reek:TooManyStatements { max_statements: 6 }
 
@@ -142,22 +66,6 @@ module Zakuro
           debug("result: #{adjusted.format}")
 
           adjusted
-        end
-
-        #
-        # 補正値を得る
-        #
-        # @return [Integer] 補正値
-        #
-        def correction_value
-          sun = correction_solar_value
-          moon = correction_moon_value
-
-          sum = sun + moon
-
-          debug("sun: #{sun}", "moon: #{moon}", "sun + moon : #{sum}")
-
-          sum
         end
 
         #
@@ -190,14 +98,6 @@ module Zakuro
           debug("[lunar]forward: #{forward}")
 
           Lunar::Value.get(remainder: remainder, forward: forward)
-        end
-
-        def add_quarter_moon_size
-          @average_remainder.add!(QUARTER)
-          @solar_location.add_quarter
-          @lunar_location.add_quarter
-
-          next_phase_index
         end
       end
     end
